+++ docs/changelog.md
@@ -19,6 +19,7 @@ category: Administration
 * test: add unit test coverage for AdditionalFieldService (#21)
 * feat: add internationalization for additional fields configuration (#21)
 * feat: use Atlassian multi select to configure addition fields (#21)
+* test: improve tests for global configuration action (#21)
 
 ### [20.08.0] - 2020-08-27